move to components/release folder
