+++ packages/ui/src/components/release/Release.module.css
@@ -111,7 +111,7 @@ ul li {
   background-color: var(--white);
   border: var(--border);
   border-radius: 0.5rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  box-shadow: 0 4px 6px var(--shadow);
   z-index: 10;
   overflow-y: hidden;
 }
@@ -145,7 +145,7 @@ ul li {
   text-align: right;
   padding-right: 1.5rem;
   font-size: 0.875rem;
-  color: #6f6f6f;
+  color: var(--description);
 }
 
 .timeline-dot {
@@ -170,9 +170,9 @@ ul li {
   width: 1rem;
   height: 1rem;
   border-radius: 9999px;
-  background-color: #5b3fff;
+  background-color: var(--primary);
   border: var(--border);
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+  box-shadow: 0 4px 6px var(--shadow);
   flex-shrink: 0;
 }
 
@@ -189,16 +189,16 @@ ul li {
 .screenshot-wrap {
   margin-bottom: 1.5rem;
   padding: 1rem;
-  background-color: #e8e4ff;
+  background-color: var(--white);
   border-radius: 1rem;
-  box-shadow: 0 1px 4px rgba(0, 0, 0, 0.03);
+  box-shadow: 0 4px 6px var(--shadow);
 }
 
 .screenshot-inner {
   border-radius: 0.75rem;
   overflow: hidden;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.06);
-  background: #fff;
+  box-shadow: 0 4px 6px var(--shadow);
+  background: var(--white);
 }
 
 .screenshot-inner img {
@@ -229,19 +229,23 @@ ul li {
   width: 0.5rem;
   height: 0.5rem;
   border-radius: 9999px;
-  background-color: #6f6f6f;
+  background-color: var(--description);
   margin-top: 0.375rem;
   flex-shrink: 0;
 }
 
 .note-text {
-  color: #6f6f6f;
+  color: var(--description);
   line-height: 1.6;
+  
 }
 
 .note-link a {
-  color: #5b3fff;
-  text-decoration: none;
+  color: var(--description);
+  text-decoration: underline;
+}
+.note-link a:hover {
+  color: var(--primary)
 }
 
 .timeline-connector {
